Corrected the collection name
The `document` collection was inappropriately named `documents` in two places. The extra `s` has been removed in those cases. To further avoid confusion with the use of the common nouns "document" and "documents" used elsewhere, the collection name has been adjusted to be `document` in each instance.

I would recommend using an alternate word instead of "document[s]" for the collection, or switching the common nouns used to something else, such as "object[s]" or "item[s]".