feat(collection): uses countDocuments, added estimatedDocumentCount
Updated `collection.count()` to point to `collection.countDocuments()` in the mongo driver. This was
added in in version 3.1 and monk has a dependency of 3.2.3 or higher of the `mongodb` package, so there should be no concern about breaking changes. The
`collection.count()` is deprecated in the `mongodb` package and suggests using
`collection.countDocuments()` instead. Documentation was updated to point to the `countDocuments`
method on the MongoDB API reference. When the `countDocuments` method was added another method was
added along with it, `estimatedDocumentCounts()`. This also adds that function to monk, the
documentation was added for it and the typings file was updated to support it as well.

resolves #299